+++ include/seqan3/alignment/pairwise/detail/type_traits.hpp
@@ -33,7 +33,7 @@
 #include <seqan3/core/detail/empty_type.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
-#include <seqan3/core/simd/simd.hpp>
+#include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/range/views/chunk.hpp>
 #include <seqan3/range/views/zip.hpp>
 #include <seqan3/utility/type_traits/function_traits.hpp>
