+++ test/performance/core/simd/view_iota_simd_benchmark.cpp
@@ -11,7 +11,7 @@
 #include <benchmark/benchmark.h>
 
 #include <seqan3/utility/simd/algorithm.hpp>
-#include <seqan3/core/simd/simd.hpp>
+#include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/core/simd/view_iota_simd.hpp>
 
 // ============================================================================
