+++ test/performance/core/simd/view_to_simd_chunk_benchmark.cpp
@@ -16,7 +16,7 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
-#include <seqan3/core/simd/simd.hpp>
+#include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/core/simd/view_to_simd.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 #include <seqan3/range/views/to.hpp>
