+++ test/snippet/utility/simd/simd.cpp
