+++ test/unit/alignment/matrix/alignment_optimum_test.cpp
@@ -10,7 +10,7 @@
 #include <type_traits>
 
 #include <seqan3/alignment/matrix/alignment_optimum.hpp>
-#include <seqan3/core/simd/simd.hpp>
+#include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/utility/simd/concept.hpp>
 
 #include <seqan3/test/simd_utility.hpp>
