+++ test/unit/alignment/matrix/detail/coordinate_matrix_simd_test.cpp
@@ -10,7 +10,7 @@
 #include <vector>
 
 #include <seqan3/alignment/matrix/detail/coordinate_matrix.hpp>
-#include <seqan3/core/simd/simd.hpp>
+#include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/test/simd_utility.hpp>
 
 #include "../../../range/iterator_test_template.hpp"
