+++ test/unit/alignment/matrix/detail/score_matrix_single_column_simd_test.cpp
@@ -10,7 +10,7 @@
 #include <vector>
 
 #include <seqan3/alignment/matrix/detail/score_matrix_single_column.hpp>
-#include <seqan3/core/simd/simd.hpp>
+#include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/test/simd_utility.hpp>
 
 #include "../../../range/iterator_test_template.hpp"
