+++ test/unit/core/simd/debug_stream_simd_test.cpp
@@ -9,7 +9,7 @@
 #include <gtest/gtest.h>
 
 #include <seqan3/core/simd/debug_stream_simd.hpp>
-#include <seqan3/core/simd/simd.hpp>
+#include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/utility/simd/algorithm.hpp>
 
 #include <iostream>
