+++ test/unit/core/simd/view_to_simd_test.cpp
@@ -19,7 +19,7 @@
 #include <seqan3/core/bit_manipulation.hpp>
 #include <seqan3/core/simd/debug_stream_simd.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
-#include <seqan3/core/simd/simd.hpp>
+#include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/core/simd/view_to_simd.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
