+++ test/unit/utility/simd/CMakeLists.txt
@@ -1,1 +1,2 @@
 seqan3_test (simd_algorithm_test.cpp)
+seqan3_test (simd_test.cpp)
