+++ test/unit/utility/simd/simd_test.cpp
@@ -8,7 +8,7 @@
 #include <gtest/gtest.h>
 
 #include <seqan3/utility/simd/concept.hpp>
-#include <seqan3/core/simd/simd.hpp>
+#include <seqan3/utility/simd/simd.hpp>
 
 TEST(simd, auto_length)
 {
